Add support for PROGRAMFILES installations
Hello! Love this script. This didn't work when running with an enterprise-installed slack client (because slack was installed into Program Files).

There's probably a more elegant way to handle this logic, but this is what I was able to get working.

Thanks again! I know this is a bit of a corner-case